fix 4

+++ include/task3.h
@@ -1,5 +1,5 @@
 
-template<class T, size_t size>
+template<class T, int size>
 void map(T (&arr)[size], T(*change)(T x)) {
 	for (int i = 0; i < size; i++) {
 		arr[i] = change(arr[i]);
